+++ Magic/CHANGELOG.md
@@ -59,6 +59,7 @@
  - Fix NPCs sometimes not respawning
  - Add simpler `slot` parameter to Wear action, boots,helmet,chesplate or leggings
  - Add `direction_count` and `source_location` parameters to CheckBlock action, more control over where it looks
+ - Add "rain" option to Weather action
 
 # 8.1.2